Flashlight working.
Technically, we should only show the light button is at least one camera
supports a flash.  The alternative to hide the light button when the
current camera doesn't support flash (or we are in image mode), but this
makes the buttons bounce back and forth.  We could alternately disable
the button.